+++ sapling-crypto/src/circuit.rs
@@ -1,6 +1,6 @@
 //! The Sapling circuits.
 
-use alloc::vec::Vec;
+use alloc::{sync::Arc, vec::Vec};
 use core::fmt;
 use corez::io;
 
@@ -570,29 +570,55 @@ impl SpendParameters {
 
     /// Returns the verifying key for the Sapling Spend circuit.
     pub fn verifying_key(&self) -> SpendVerifyingKey {
-        SpendVerifyingKey(self.0.vk.clone())
+        SpendVerifyingKey::new(self.0.vk.clone())
     }
 
-    /// Returns the verifying key for the Sapling Spend circuit, with precomputations
-    /// optimized for verifying individual proofs.
+    /// Returns the verifying key for the Sapling Spend circuit, with
+    /// precomputations optimized for verifying individual proofs.
     pub fn prepared_verifying_key(&self) -> PreparedSpendVerifyingKey {
-        PreparedSpendVerifyingKey(groth16::prepare_verifying_key(&self.0.vk))
+        PreparedSpendVerifyingKey(Arc::new(groth16::prepare_verifying_key(&self.0.vk)))
     }
 }
 
 /// The verifying key for the Sapling Spend circuit.
-pub struct SpendVerifyingKey(pub(crate) groth16::VerifyingKey<Bls12>);
+pub struct SpendVerifyingKey {
+    verifying_key: groth16::VerifyingKey<Bls12>,
+    prepared_verifying_key: Arc<groth16::PreparedVerifyingKey<Bls12>>,
+}
 
 impl SpendVerifyingKey {
-    /// Performs precomputations optimized for verifying individual proofs.
+    fn new(verifying_key: groth16::VerifyingKey<Bls12>) -> Self {
+        let prepared_verifying_key = Arc::new(groth16::prepare_verifying_key(&verifying_key));
+        Self {
+            verifying_key,
+            prepared_verifying_key,
+        }
+    }
+
+    /// Returns this verifying key with precomputations optimized for verifying
+    /// individual proofs.
     pub fn prepare(&self) -> PreparedSpendVerifyingKey {
-        PreparedSpendVerifyingKey(groth16::prepare_verifying_key(&self.0))
+        PreparedSpendVerifyingKey(Arc::clone(&self.prepared_verifying_key))
+    }
+
+    pub(super) fn batch_verifying_key(&self) -> &groth16::VerifyingKey<Bls12> {
+        &self.verifying_key
+    }
+
+    pub(super) fn prepared_verifying_key(&self) -> &groth16::PreparedVerifyingKey<Bls12> {
+        &self.prepared_verifying_key
     }
 }
 
-/// The verifying key for the Sapling Spend circuit, with precomputations optimized for
-/// verifying individual proofs.
-pub struct PreparedSpendVerifyingKey(pub(crate) groth16::PreparedVerifyingKey<Bls12>);
+/// The verifying key for the Sapling Spend circuit, with precomputations
+/// optimized for verifying individual proofs.
+pub struct PreparedSpendVerifyingKey(Arc<groth16::PreparedVerifyingKey<Bls12>>);
+
+impl PreparedSpendVerifyingKey {
+    pub(super) fn as_inner(&self) -> &groth16::PreparedVerifyingKey<Bls12> {
+        &self.0
+    }
+}
 
 /// The parameters for the Sapling Output circuit.
 pub struct OutputParameters(pub(crate) groth16::Parameters<Bls12>);
@@ -608,29 +634,55 @@ impl OutputParameters {
 
     /// Returns the verifying key for the Sapling Output circuit.
     pub fn verifying_key(&self) -> OutputVerifyingKey {
-        OutputVerifyingKey(self.0.vk.clone())
+        OutputVerifyingKey::new(self.0.vk.clone())
     }
 
-    /// Returns the verifying key for the Sapling Output circuit, with precomputations
-    /// optimized for verifying individual proofs.
+    /// Returns the verifying key for the Sapling Output circuit, with
+    /// precomputations optimized for verifying individual proofs.
     pub fn prepared_verifying_key(&self) -> PreparedOutputVerifyingKey {
-        PreparedOutputVerifyingKey(groth16::prepare_verifying_key(&self.0.vk))
+        PreparedOutputVerifyingKey(Arc::new(groth16::prepare_verifying_key(&self.0.vk)))
     }
 }
 
 /// The verifying key for the Sapling Output circuit.
-pub struct OutputVerifyingKey(pub(crate) groth16::VerifyingKey<Bls12>);
+pub struct OutputVerifyingKey {
+    verifying_key: groth16::VerifyingKey<Bls12>,
+    prepared_verifying_key: Arc<groth16::PreparedVerifyingKey<Bls12>>,
+}
 
 impl OutputVerifyingKey {
-    /// Performs precomputations optimized for verifying individual proofs.
+    fn new(verifying_key: groth16::VerifyingKey<Bls12>) -> Self {
+        let prepared_verifying_key = Arc::new(groth16::prepare_verifying_key(&verifying_key));
+        Self {
+            verifying_key,
+            prepared_verifying_key,
+        }
+    }
+
+    /// Returns this verifying key with precomputations optimized for verifying
+    /// individual proofs.
     pub fn prepare(&self) -> PreparedOutputVerifyingKey {
-        PreparedOutputVerifyingKey(groth16::prepare_verifying_key(&self.0))
+        PreparedOutputVerifyingKey(Arc::clone(&self.prepared_verifying_key))
+    }
+
+    pub(super) fn batch_verifying_key(&self) -> &groth16::VerifyingKey<Bls12> {
+        &self.verifying_key
+    }
+
+    pub(super) fn prepared_verifying_key(&self) -> &groth16::PreparedVerifyingKey<Bls12> {
+        &self.prepared_verifying_key
     }
 }
 
-/// The verifying key for the Sapling Output circuit, with precomputations optimized for
-/// verifying individual proofs.
-pub struct PreparedOutputVerifyingKey(pub(crate) groth16::PreparedVerifyingKey<Bls12>);
+/// The verifying key for the Sapling Output circuit, with precomputations
+/// optimized for verifying individual proofs.
+pub struct PreparedOutputVerifyingKey(Arc<groth16::PreparedVerifyingKey<Bls12>>);
+
+impl PreparedOutputVerifyingKey {
+    pub(super) fn as_inner(&self) -> &groth16::PreparedVerifyingKey<Bls12> {
+        &self.0
+    }
+}
 
 #[test]
 fn test_input_circuit_with_bls12_381() {